chore: refactor ssr
