+++ packages/ssr/src/client/common/topic.js
