+++ packages/ssr/src/client/entry/author-profile.js
@@ -1,0 +1,5 @@
+const authorProfile = require("../common/author-profile");
+const makeArticleUrl = require("../../lib/make-url");
+const defaultMapProfileToConfig = require("../../lib/make-ad-config");
+
+authorProfile(makeArticleUrl, defaultMapProfileToConfig);
