+++ packages/ssr/src/lib/make-ad-config.js
@@ -1,0 +1,8 @@
+module.exports = () => ({
+  adUnit: '',
+  biddersConfig: {},
+  bidderSlots: [],
+  networkId: '',
+  pageTargeting: {},
+  slotTargeting: {}
+});
